move settings_tab.ts; refactor: extract createSnippetsEditor into own function
