+++ src/main.ts
@@ -1,6 +1,6 @@
 import { Plugin, Notice } from "obsidian";
 import { LatexSuiteSettings, DEFAULT_SETTINGS } from "./settings";
-import { LatexSuiteSettingTab } from "./settings_tab";
+import { LatexSuiteSettingTab } from "./ui/settings_tab";
 
 import { ctxAtViewPos } from "./snippets/context";
